Problems now return a ulong. Added P7.

+++ Solutions/Problems/IProblem.cs
@@ -2,6 +2,6 @@ namespace Solutions
 {
     public interface IProblem
     {
-        int Run(ulong targetNumber);
+        ulong Run(ulong targetNumber);
     }
 }
