+++ Solutions/Problems/Problem3.cs
@@ -6,13 +6,13 @@ namespace Solutions
 {
     public class Problem3 : IProblem
     {
-        public int Run(ulong targetNumber)
+        public ulong Run(ulong targetNumber)
         {
             var primes = new PrimeNumbers();
 
             var primeFactors = primes.GetPrimeFactors(targetNumber);
 
-            return (int)primeFactors.Max();
+            return primeFactors.Max();
         }
